Change assertions to RuntimeErrors
Assertions shouldn't be used to raise exceptions in production code.
Instead they're replaced by RuntimeErrors.

+++ ltdconveyor/s3copy.py
@@ -79,6 +79,8 @@ def copy_dir(bucket_name, src_path, dest_path,
     ------
     ltdconveyor.exceptions.S3Error
         Thrown by any unexpected faults from the S3 API.
+    RuntimeError
+        Thrown when the source and destination directories are the same.
     """
     if not src_path.endswith('/'):
         src_path += '/'
@@ -87,8 +89,14 @@ def copy_dir(bucket_name, src_path, dest_path,
 
     # Ensure the src_path and dest_path don't contain each other
     common_prefix = os.path.commonprefix([src_path, dest_path])
-    assert common_prefix != src_path
-    assert common_prefix != dest_path
+    if common_prefix == src_path:
+        msg = 'Common prefix {0} is same as source dir {1}'.format(
+            common_prefix, src_path)
+        raise RuntimeError(msg)
+    if common_prefix == dest_path:
+        msg = 'Common prefix {0} is same as dest dir {1}'.format(
+            common_prefix, dest_path)
+        raise RuntimeError(msg)
 
     # Delete any existing objects in the destination
     delete_dir(bucket_name, dest_path,
